update dependencies and files needed for code generation
