+++ bundles/org.openhab.ui.cometvisu/pom.xml
@@ -61,12 +61,12 @@
     <dependency>
       <groupId>org.rrd4j</groupId>
       <artifactId>rrd4j</artifactId>
-      <version>3.4</version>
+      <version>3.8.2</version>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-annotations</artifactId>
-      <version>2.9.10</version>
+      <version>2.14.0-rc2</version>
     </dependency>
   </dependencies>
 
@@ -107,7 +107,7 @@
             </goals>
             <configuration>
               <!-- The package of your generated sources -->
-              <packageName>org.openhab.ui.cometvisu.internal.backend.model.config</packageName>
+              <packageName>org.openhab.ui.cometvisu.internal.backend.model.config.pure</packageName>
               <sources>
                 <source>src/main/resources/visu_config.xsd</source>
               </sources>
