+++ bundles/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/model/config/pure/CDataStatus.java
@@ -10,49 +10,23 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.ui.cometvisu.internal.backend.model.config;
+package org.openhab.ui.cometvisu.internal.backend.model.config.pure;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlType;
-import javax.xml.bind.annotation.XmlValue;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
 /**
- * Adapter for marshaling label with CDATA content
+ * Adapter for marshaling status with CDATA content
  *
  * @author Tobias Bräutigam - Initial contribution
  *
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "label", propOrder = { "value" })
-public class CDataLabel extends Label {
+@XmlType(name = "status", propOrder = { "value" })
+public class CDataStatus extends Status {
 
     @XmlJavaTypeAdapter(AdapterCDATA.class)
-    @XmlValue
     protected String value;
-
-    /**
-     * Gets the value of the value property.
-     *
-     * @return
-     *         possible object is
-     *         {@link String }
-     *
-     */
-    public String getValue() {
-        return value;
-    }
-
-    /**
-     * Sets the value of the value property.
-     *
-     * @param value
-     *            allowed object is
-     *            {@link String }
-     *
-     */
-    public void setValue(String value) {
-        this.value = value;
-    }
 }
