+++ bundles/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/sitemap/ConfigHelper.java
@@ -38,36 +38,36 @@ import org.openhab.core.types.TypeParser;
 import org.openhab.core.ui.icon.IconProvider;
 import org.openhab.core.ui.icon.IconSet.Format;
 import org.openhab.ui.cometvisu.internal.Config;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Address;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Button;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Buttons;
-import org.openhab.ui.cometvisu.internal.backend.model.config.CDataStatus;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Entry;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Group;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Icon;
-import org.openhab.ui.cometvisu.internal.backend.model.config.IconDefinition;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Icons;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Info;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Label;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Layout;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Line;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Mapping;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Mappings;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Multitrigger;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Navbar;
-import org.openhab.ui.cometvisu.internal.backend.model.config.NavbarPositionType;
-import org.openhab.ui.cometvisu.internal.backend.model.config.ObjectFactory;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Page;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Pagejump;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Pages;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Plugin;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Plugins;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Statusbar;
-import org.openhab.ui.cometvisu.internal.backend.model.config.StylingEntry;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Stylings;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Text;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Trigger;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Widgetinfo;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Address;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Button;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Buttons;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.CDataStatus;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Entry;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Group;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Icon;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.IconDefinition;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Icons;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Info;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Label;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Layout;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Line;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Mapping;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Mappings;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Multitrigger;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Navbar;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.NavbarPositionType;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.ObjectFactory;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Page;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Pagejump;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Pages;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Plugin;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Plugins;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Statusbar;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.StylingEntry;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Stylings;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Text;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Trigger;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Widgetinfo;
 import org.openhab.ui.cometvisu.internal.servlet.CometVisuApp;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
