+++ bundles/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/sitemap/VisuConfig.java
@@ -40,28 +40,28 @@ import org.openhab.core.model.sitemap.sitemap.Sitemap;
 import org.openhab.core.model.sitemap.sitemap.Webview;
 import org.openhab.core.model.sitemap.sitemap.Widget;
 import org.openhab.ui.cometvisu.internal.Config;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Address;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Colorchooser;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Diagram;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Group;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Image;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Info;
-import org.openhab.ui.cometvisu.internal.backend.model.config.LibVersion;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Mapping;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Meta;
-import org.openhab.ui.cometvisu.internal.backend.model.config.NavbarPositionType;
-import org.openhab.ui.cometvisu.internal.backend.model.config.ObjectFactory;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Page;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Pages;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Plugin;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Rrd;
-import org.openhab.ui.cometvisu.internal.backend.model.config.SchemaPages;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Slide;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Switch;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Text;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Trigger;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Video;
-import org.openhab.ui.cometvisu.internal.backend.model.config.Web;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Address;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Colorchooser;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Diagram;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Group;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Image;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Info;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.LibVersion;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Mapping;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Meta;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.NavbarPositionType;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.ObjectFactory;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Page;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Pages;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Plugin;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Rrd;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.SchemaPages;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Slide;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Switch;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Text;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Trigger;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Video;
+import org.openhab.ui.cometvisu.internal.backend.model.config.pure.Web;
 import org.openhab.ui.cometvisu.internal.backend.sitemap.ConfigHelper.Transform;
 import org.openhab.ui.cometvisu.internal.servlet.CometVisuApp;
 import org.rrd4j.ConsolFun;
